+++ src/components/commons/HomeContent/HomeContent.tsx
@@ -11,9 +11,7 @@ const HomeContent = () => {
     actionCreator,
     dispatch
   );
-  const stateGenres = useSelector((state: any) => state.home.genres);
-  const stateChart = useSelector((state: any) => state.home.chart);
-  const stateLast = useSelector((state: any) => state.home.last);
+  const { last, genres, chart } = useSelector((state: any) => state.home);
 
   useEffect(() => {
     getGenres();
@@ -23,18 +21,24 @@ const HomeContent = () => {
 
   return (
     <div className={styles.container}>
-      <div className={styles.section}>
-        <h1>Recently played</h1>
-        <CardContainer content={stateLast.slice(0, 5)} />
-      </div>
-      <div className={styles.section}>
-        <h1>Genres</h1>
-        <CardContainer content={stateGenres} />
-      </div>
-      <div className={styles.section}>
-        <h1>Most popular</h1>
-        <CardContainer content={stateChart.slice(0, 5)} />
-      </div>
+      { last.length > 0 &&
+        <div className={styles.section}>
+          <h1>Recently played</h1>
+          <CardContainer content={last.slice(0, 5)} />
+        </div>
+      }
+      { genres.length > 0 &&
+        <div className={styles.section}>
+          <h1>Genres</h1>
+          <CardContainer content={genres} />
+        </div>
+      }
+      { chart.length > 0 &&
+        <div className={styles.section}>
+          <h1>Most popular</h1>
+          <CardContainer content={chart.slice(0, 5)} />
+        </div>
+      }
     </div>
   );
 };
