+++ src/components/views/Home/Home.module.css
@@ -11,6 +11,7 @@
   grid-column: 2 / 3;
   grid-row: 1 / 2;
   border-bottom: 1px solid var(--gray-light);
+  z-index: 2;
 }
 .container > div:nth-child(2) {
   grid-column: 1 / 2;
@@ -25,6 +26,7 @@
   /* hidde scrollbar firefox */
   -ms-overflow-style: none;
   scrollbar-width: none;
+  z-index: 1;
 }
   /* hidde scrollbar chrome */
 .container > div:nth-child(3)::-webkit-scrollbar {
